+++ source/server/BUILD
@@ -9,9 +9,18 @@ licenses(["notice"])  # Apache 2
 envoy_package()
 
 envoy_cc_library(
-    name = "common_lib",
-    srcs = ["common.cc"],
-    hdrs = ["common.h"],
+    name = "well_known_headers_lib",
+    hdrs = ["well_known_headers.h"],
+    repository = "@envoy",
+    deps = [
+        "@envoy//source/common/singleton:const_singleton_with_external_headers",
+    ],
+)
+
+envoy_cc_library(
+    name = "configuration_lib",
+    srcs = ["configuration.cc"],
+    hdrs = ["configuration.h"],
     repository = "@envoy",
     deps = [
         "//api/server:response_options_proto_cc_proto",
@@ -27,7 +36,8 @@ envoy_cc_library(
     hdrs = ["http_test_server_filter.h"],
     repository = "@envoy",
     deps = [
-        ":common_lib",
+        ":configuration_lib",
+        ":well_known_headers_lib",
         "//api/server:response_options_proto_cc_proto",
         "@envoy//source/exe:envoy_common_lib_with_external_headers",
     ],
@@ -39,7 +49,8 @@ envoy_cc_library(
     hdrs = ["http_dynamic_delay_filter.h"],
     repository = "@envoy",
     deps = [
-        ":common_lib",
+        ":configuration_lib",
+        ":well_known_headers_lib",
         "//api/server:response_options_proto_cc_proto",
         "@envoy//source/exe:envoy_common_lib_with_external_headers",
         "@envoy//source/extensions/filters/http/fault:fault_filter_lib",
